HTTP Early Hints.
The change implements parsing Early Hints response from the proxied server
and forwarding it to the client.  Also, an Early Hints response could be
sent to the client by the upstream module before choosing a proxied server.
This happens if there are headers to send, which can be configured or added
by third-party modules using the early hint filters.

A new parameter "early_hint" is added to the "add_header" directive.
The parameter instructs nginx to send this header as part of Early Hints
response.  The header is also sent in the main response.

Example:

    add_header X-Foo foo early_hint;
    proxy_pass http://bar.example.com;