Added "keepalive_min_timeout" directive.
The directive sets a timeout during which a keepalive connection will
not be closed by nginx.  Previously, it could be closed at any time if
a connection was needed to be reused for another client, or when the
worker was gracefully shutting down.

The change addresses two problems with HTTP keepalive.  First, it allows
clients which send multiple requests over the same connection without
delay or with a small delay between them, to avoid receiving a TCP RST
in response to one of them.  After starting graceful worker shutdown,
the client receives "Connection: close" to indicate that the connection
is going to be closed.  Second, it addresses the TCP reset problem
described in RFC 9112, Section 9.6, when the last sent HTTP response
could be damaged by a followup TCP RST.  Both problems are especially
important for non-idempotent requests, which cannot be retried by client.

The change does not provide a complete solution against unexpected
connection closures.  A TCP connection can still be closed due to
network issues.  Also, a keepalive connection is closed by nginx after
keepalive_timeuot.  Normally, a client should be prepared to handle
such closures, as mentioned in RFC 9112, Section 9.5.

It is not recommended to set keepalive_min_timeout to large values as
this can introduce an additional delay during worker process shutdown
and may restrict nginx from effective connection reuse.